requirements.txt
gunicorn
Flask==0.12.2
Flask-Cors==3.0.2
Keras==2.1.5
matplotlib==2.0.2
numpy==1.12.1
pandas==0.20.1
python-telegram-bot==11.1.0
Quandl==3.4.5
scikit-learn==0.18.1
tensorflow==1.2.1
textblob==0.15.2
tweepy==3.7.0
h5py==2.8.0
nsepy
beautifulsoup4==4.4.1
bs4==0.0.1
lxml==3.8.0